+++ src/process.cpp
@@ -7,6 +7,7 @@
 #include <Rinterface.h> // R_CStackLimit
 #endif
 #include "Rcompatibility.h"
+#include "R2cpp.h"
 
 
 // STL
@@ -26,12 +27,13 @@
 #define CONCURRENTFILES 3
 #define NESTED 4
 
-SEXP process(SEXP sexppipeline, SEXP sexpprogrss, SEXP sexpncpu, SEXP sexpmode, SEXP sexpverbose)
+SEXP process(SEXP sexppipeline, SEXP args)
 {
-  int ncpu = Rf_asInteger(sexpncpu);
-  int mode = Rf_asInteger(sexpmode);
-  bool progrss = Rf_asLogical(sexpprogrss);
-  bool verbose = Rf_asLogical(sexpverbose);
+  int ncpu = get_element_as_int(args, "ncores");
+  int strategy = get_element_as_int(args, "strategy");
+  bool progrss = get_element_as_bool(args, "progress");
+  bool verbose = get_element_as_bool(args, "verbose");
+  double chunk_size = get_element_as_double(args, "chunk_size");
 
   // Check some multithreading stuff
   if (ncpu > available_threads())
@@ -41,12 +43,12 @@ SEXP process(SEXP sexppipeline, SEXP sexpprogrss, SEXP sexpncpu, SEXP sexpmode, 
   }
   int ncpu_outer_loop = 1; // concurrent files
   int ncpu_inner_loops = 1; // concurrent points
-  if (mode == CONCURRENTPOINTS) ncpu_inner_loops = ncpu;
-  if (mode == CONCURRENTFILES) ncpu_outer_loop = ncpu;
-  if (mode == NESTED)
+  if (strategy == CONCURRENTPOINTS) ncpu_inner_loops = ncpu;
+  if (strategy == CONCURRENTFILES) ncpu_outer_loop = ncpu;
+  if (strategy == NESTED)
   {
     ncpu_outer_loop = ncpu;
-    ncpu_inner_loops = INTEGER(sexpncpu)[1];
+    ncpu_inner_loops = get_element_as_vint(args, "ncores")[1];
   }
   if (ncpu_outer_loop > 1 && ncpu_inner_loops > 1) omp_set_max_active_levels(2); // nested
 
@@ -66,7 +68,12 @@ SEXP process(SEXP sexppipeline, SEXP sexpprogrss, SEXP sexpncpu, SEXP sexpmode, 
     bool is_parallelizable = pipeline.is_parallelizable();  // concurrent-files
 
     LAScatalog* lascatalog = pipeline.get_catalog(); // the pipeline owns the catalog
-    lascatalog->set_chunk_size(0);                   // currently only 0 is supported
+
+    if (!lascatalog->set_chunk_size(chunk_size))
+    {
+      throw last_error;
+    }
+
     int n = lascatalog->get_number_chunks();
 
     // Check some multi-threading stuff