+++ src/zzz.cpp
@@ -4,7 +4,7 @@
 #include <R_ext/Rdynload.h>
 #include <R_ext/Visibility.h>
 
-SEXP process(SEXP, SEXP, SEXP, SEXP, SEXP);
+SEXP process(SEXP, SEXP);
 SEXP lasfilterusage();
 SEXP lastransformusage();
 SEXP get_pipeline_info(SEXP);
@@ -12,7 +12,8 @@ SEXP is_indexed(SEXP sexpfile);
 int available_threads();
 bool has_omp_support();
 
-extern "C" SEXP C_process(SEXP pipeline, SEXP progrss, SEXP ncpu, SEXP nfiles, SEXP verbose) { return process(pipeline, progrss, ncpu, nfiles, verbose); }
+
+extern "C" SEXP C_process(SEXP pipeline, SEXP args) { return process(pipeline, args); }
 extern "C" SEXP C_lasfilterusage() { return lasfilterusage(); }
 extern "C" SEXP C_lastransformusage() { return lastransformusage(); }
 extern "C" SEXP C_get_pipeline_info(SEXP pipeline) { return get_pipeline_info(pipeline); }
@@ -24,7 +25,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"C_lasfilterusage", (DL_FUNC) &C_lasfilterusage, 0},
   {"C_lastransformusage", (DL_FUNC) &C_lastransformusage, 0},
   {"C_get_pipeline_info", (DL_FUNC) &C_get_pipeline_info, 1},
-  {"C_process", (DL_FUNC) &C_process, 5},
+  {"C_process", (DL_FUNC) &C_process, 2},
   {"C_available_threads", (DL_FUNC) &C_available_threads, 0},
   {"C_has_omp_support", (DL_FUNC) &C_has_omp_support, 0},
   {"C_is_indexed", (DL_FUNC) &C_is_indexed, 1},