+++ tests/testthat/test-classify_with_ivf.R
@@ -1,7 +1,6 @@
 test_that("classify noise with ivf works",
 {
   f <- system.file("extdata", "Topography.las", package="lasR")
-  read = reader(f)
   class = classify_isolated_points()
-  expect_error(processor(read + class), NA)
+  expect_error(exec(class, f), NA)
 })